Move grpc to primary http server
