+++ grpc/tx.go
@@ -1,4 +1,4 @@
-package grpc
+package junogrpc
 
 import (
 	"errors"
